+++ Packages/com.alelievr.mixture/Runtime/Nodes/Utils/RandomColorNode.cs
@@ -4,6 +4,8 @@ using UnityEngine;
 using GraphProcessor;
 using System.Linq;
 using UnityEngine.Rendering;
+using System;
+using Random = System.Random;
 
 namespace Mixture
 {
@@ -12,7 +14,7 @@ Output a random color based on the HSV parameters.
 ")]
 
 	[System.Serializable, NodeMenuItem("Math/Random Color")]
-	public class RandomColorNode : MixtureNode
+	public class RandomColorNode : MixtureNode, IRealtimeReset
 	{
 		public override bool hasSettings => false;
 
@@ -31,13 +33,22 @@ Output a random color based on the HSV parameters.
 
 		public override string name => "Random Color";
 
+		Random random;
+
+		protected override void Enable()
+			=> InitSeed();
+
 		protected override bool ProcessNode(CommandBuffer cmd)
 		{
-			var oldState = Random.state;
-			Random.InitState(seed);
-			r = Random.ColorHSV(minHue, maxHue, minSat, maxSat, minValue, maxValue);
-			Random.state = oldState;
+			float hue = Mathf.Lerp(minHue, maxHue, (float)random.NextDouble());
+			float sat = Mathf.Lerp(minSat, maxSat, (float)random.NextDouble());
+			float val = Mathf.Lerp(minValue, maxValue, (float)random.NextDouble());
+			r = Color.HSVToRGB(hue, sat, val);
 			return true;
 		}
+
+		void InitSeed() => random = new Random(seed);
+
+		public void RealtimeReset() => InitSeed();
 	}
 }